[core] TileID coordinates now have double precision
This is a prior step towards removing TileCoordinate entirely.

+++ src/mbgl/map/source.cpp
@@ -168,7 +168,7 @@ void Source::load() {
 void Source::updateMatrices(const mat4 &projMatrix, const TransformState &transform) {
     for (const auto& pair : tiles) {
         Tile &tile = *pair.second;
-        transform.matrixFor(tile.matrix, tile.id, std::min(static_cast<int8_t>(info->maxZoom), tile.id.z));
+        transform.matrixFor(tile.matrix, tile.id, std::min(info->maxZoom, tile.id.z));
         matrix::multiply(tile.matrix, projMatrix, tile.matrix);
     }
 }
